updated tests

+++ src/main/java/com/qa/main/TemperatureConverter.java
@@ -2,29 +2,28 @@ package com.qa.main;
 
 public class TemperatureConverter {
 	
-	public float convertFahrenheitToCelsius(int fahrenheit) {
-        return (5 / 9) * (fahrenheit - 32);
+	public double convertFahrenheitToCelsius(double fahrenheit) {
+        return ((double) (5 / 9) * (fahrenheit - 32));
     }
 
-    public float convertCelsiusToFahrenheit(int celsius) {
+    public double convertCelsiusToFahrenheit(double celsius) {
         return (9 / 5) * (celsius) + 32;
     }
 
-    public float convertKelvinToCelsius(int kelvin) {
+    public double convertKelvinToCelsius(double kelvin) {
         return (kelvin - 273);
     }
 
-    public float convertCelsiusToKelvin(int celsius) {
+    public double convertCelsiusToKelvin(double celsius) {
         return (celsius + 273);
     }
 
-    public float convertKelvinToFahrenheit(int kelvin) {
+    public double convertKelvinToFahrenheit(double kelvin) {
         return (9 / 5) * (kelvin - 273) + 32;
     }
 
-    public float convertFahrenheitToKelvin(int fahrenheit) {
+    public double convertFahrenheitToKelvin(double fahrenheit) {
         return (5 / 9) * (fahrenheit - 32) + 273;
     }
-
 
 }
